consolidate settings module
