+++ config/__init__.py
@@ -1,10 +1,25 @@
 """Convenient access to application settings and constants."""
 
-from .settings import ETLConstants, Settings, settings, parse_database_name
+from .settings import (
+    ETLConstants,
+    Settings,
+    settings,
+    parse_database_name,
+    SecretManager,
+    ConfigurationManager,
+    get_settings,
+    get_secure_settings,
+    migrate_existing_configuration,
+)
 
 __all__ = [
     "settings",
     "Settings",
     "parse_database_name",
     "ETLConstants",
+    "SecretManager",
+    "ConfigurationManager",
+    "get_settings",
+    "get_secure_settings",
+    "migrate_existing_configuration",
 ]
